refactor(agenda): déplacer la logique d'auth dans le module
Les helpers requireAgendaView/Manage/Qualify et isProtocoleMember
appartiennent au module agenda, pas à src/lib/auth.ts (lib globale).

- Crée src/modules/agenda/auth.ts avec les 4 fonctions
- Les exporte via l'index public du module
- Met à jour les 11 fichiers src/app/ pour importer depuis @/modules/agenda
- Supprime les 4 fonctions de src/lib/auth.ts

lint:boundaries : 0 violation — 326 modules, 1167 dépendances vérifiées.
