+++ src/modules/agenda/index.ts
@@ -1,5 +1,7 @@
 import { defineModule } from "@/core/module-registry";
 
+export { isProtocoleMember, requireAgendaView, requireAgendaManage, requireAgendaQualify } from "./auth";
+
 /**
  * Module agenda — Agenda pastoral.
  *
